chore(files): Restructure app directory and update dependencies
The project restructures the `app` directory:
- Moves the `routes/api` and `routes/site` directories up one level.
- Adds a new `cache` directory to the `app` directory.
- Updates references to the moved files in the README.md file.

This change aims to improve the organization of the project by grouping
related files together logically within the `app` directory.
Additionally, it enhances clarity for developers working on the project.
Dependencies were updated in `requirements.txt` to ensure compatibility with
the restructured project layout.